My name program

+++ C5/C5.c
@@ -5,7 +5,7 @@
 int main(void)
 {
     int r=2; 
-int Vol= (4*M_PI*r*r*r)/3;
-int Area=4*M_PI*r*r;
-printf("The volume of the sphere is %d\n and the surface area is%d",Vol,Area);
+float Vol= (4*M_PI*r*r*r)/3;
+float Area=4*M_PI*r*r;
+printf("The volume of the sphere is %-3f\n and the surface area is%3f",Vol,Area);
 }
